% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMA.R
\name{Bayesian_Model_Averaging}
\alias{Bayesian_Model_Averaging}
\title{Bayesian Model Averaging}
\usage{
Bayesian_Model_Averaging(Model_List, Target)
}
\arguments{
\item{Model_List}{A list of numeric probabilities vector}

\item{Target}{Ground truth (correct) 0-1 labels vector}
}
\value{
Ensemble Weights
}
\description{
Computing Ensemble Weights using Bayesian Model Averaging.
}
