% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMC.R
\name{Bayesian_Model_Combination}
\alias{Bayesian_Model_Combination}
\title{Bayesian Model Combination Function}
\usage{
Bayesian_Model_Combination(Model_List, Target, Sample_Size, Sample_Time,
  Prior = NULL, Verbose = FALSE)
}
\arguments{
\item{Model_List}{A list of numeric probabilities vector}

\item{Target}{Ground truth (correct) 0-1 labels vector}

\item{Sample_Size}{Number of weights to be random sampled}

\item{Sample_Time}{Number of times for weights to be random sampled}

\item{Prior}{Prior for the ensemble weights}

\item{Verbose}{Whether or not to print progress.}
}
\value{
Ensemble Weights
}
\description{
Computing Ensemble Weights using Bayesian Model Combination iteratively.
}
