% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMC.R
\name{Bayesian_Model_Combination_Helper}
\alias{Bayesian_Model_Combination_Helper}
\title{Bayesian Model Combination Helper Function}
\usage{
Bayesian_Model_Combination_Helper(Model_List, Target, Prior = NULL,
  Sample_Size)
}
\arguments{
\item{Model_List}{A list of numeric probabilities vector}

\item{Target}{Ground truth (correct) 0-1 labels vector}

\item{Prior}{Prior for the ensemble weights}

\item{Sample_Size}{Number of weights to be random sampled}
}
\value{
Ensemble Weights
}
\description{
Computing Ensemble Weights in one round for Bayesian Model Combination.
}
\keyword{internal}
