% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMC.R
\name{Bayesian_Model_Combination_Naive}
\alias{Bayesian_Model_Combination_Naive}
\title{Bayesian Model Combination Naive}
\usage{
Bayesian_Model_Combination_Naive(Model_List, Target, Sample_Size)
}
\arguments{
\item{Model_List}{A list of numeric probabilities vector}

\item{Target}{Ground truth (correct) 0-1 labels vector}

\item{Sample_Size}{Number of weights to be random sampled}
}
\value{
Ensemble Weights
}
\description{
Computing Ensemble Weights using Bayesian Model Combination Naive Version.
}
