% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cornheight.R
\docType{data}
\name{cornheight}
\alias{cornheight}
\title{Heights of offspring of Zea mays}
\format{
A data frame with 15 pairs of observations and three variables:
\describe{
\item{Pair}{Pair number, a numeric vector}
\item{Cross}{Height of cross-fertilized plants, a numeric vector}
\item{Self}{Height of self-fertilized plants, a numeric vector}
}
}
\source{
Fisher, R.A. (1935) \emph{The Design of Experiments}. Oliver and
Boyd. Edinburgh.
}
\usage{
cornheight
}
\description{
A dataset containing the height of 15 pairs of the offspring of maize plants.
}
\details{
{
Darwin took 15 pairs of \emph{Zea Mays} plants, where within each pair the
two plants were of exactly the same age, were subjected from the first to
last to the same conditions, and were descended from the same parents. One
individual in each pair was cross-fertilized and the other was
self-fertilized. The heights of offspring were then measured to the nearest
eighth of an inch. This data set was used by Fisher (1935) to introduce the
idea of the randomization test with paired comparison data.
}
}
\keyword{datasets}
