% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackals.R
\docType{data}
\name{jackals}
\alias{jackals}
\title{Mandible Lengths of Male and Female Golden Jackals.}
\format{
A data frame with 20 rows and 2 variables:
\describe{
\item{Sex}{Sex of the golden jackal}
\item{Mand.length}{Mandible length, in millimeters}
...
}
}
\source{
Higham, C.F.W., Kijngam, A. and Manly, B.F.J. (1980).
An analysis of prehistoric canid remains from Thailand. Journal
of Archaeological Science 7: 149-165.
}
\usage{
jackals
}
\description{
A dataset containing mandible lenghts in millimiters for male and female
golden jackals (\emph{Canis aureus}) for 10 of each sex in the collection in
the British Museum of Natural History in London.
}
\keyword{datasets}
