% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{Optimal_Accuracy}
\alias{Optimal_Accuracy}
\title{Optimal Accuracy Computing}
\usage{
Optimal_Accuracy(pred, label)
}
\arguments{
\item{pred}{Predicted labels vector, as returned by a classifier}

\item{label}{Ground truth (correct) 0-1 labels vector}
}
\value{
AUC
}
\description{
Computing AUC
}
\keyword{internal}
