% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbmc-package.R
\docType{package}
\name{rbmc-package}
\alias{rbmc}
\alias{rbmc-package}
\title{rbmc: A package to accompany the book "Randomization, Bootstrap and Monte Carlo methods in Biology. Fourth Edition"}
\description{
Contains functions from: Manly, B.F.J. and Navarro-Alberto, J.A. Randomization, Bootstrap and Monte Carlo Methods in Biology. 4th Edition. CRC/Taylor & Francis, Boca Raton, FL.
}
\details{
The rbmc package provides procedures for:
One and two sample randomization tests, Randomization-based ANOVA,
Regression Analysis, Spatial Data Analysis, Time Series Analysis,
Survival and Growth Data Analysis, Analysis of Species Co-occurrences and
Bayesian methods
}
\author{
\strong{Maintainer}: Jorge A. Navarro-Alberto \email{ganava4@gmail.com}

Authors:
\itemize{
  \item Bryan F.J. Manly
}

}
\keyword{internal}
