% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RT1SAMP.R
\name{RT1SAMP}
\alias{RT1SAMP}
\title{One-sample or Paired comparison randomization test}
\usage{
RT1SAMP(
  x,
  mu = 0,
  NRAND = 4999,
  alt = "two",
  complete = FALSE,
  CI = FALSE,
  silent = FALSE,
  seed = NULL
)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{mu}{an scalar; it is the value of \code{mu} under the null hypothesis}

\item{NRAND}{numeric; the number of randomizations (permutations). It is
ignored when \code{complete=TRUE}.}

\item{alt}{a character string specifying the alternative hypothesis, must
be one of "\code{two}" (two-sided, the default), "\code{greater}" or
"\code{less}".}

\item{complete}{a logical variable; when \code{TRUE}, all possible
randomizations are enumerated. It is ignored when \code{CI=TRUE}.}

\item{CI}{a logical variable indicating whether approximate 95\\% and 99\\%
randomization confidence intervals will be calculated (\code{TRUE}) or not
(\code{FALSE}).}

\item{silent}{a logical variable indicating whether calculation results are
printed to the R console (\code{silent = FALSE}). If \code{TRUE} then
calculation results are not printed to the R console (useful for
simulations)}

\item{seed}{a single value, interpreted as an integer, or \code{NULL} (see
"Details").}
}
\value{
The function returns a \code{RT} result object (list)
}
\description{
RT1SAMP carries out the Fisher randomization test or the paired comparison
randomization test. The permutational one-sample randomization test does not
require normality of the distribution. A randomization confidence interval
for the mean can also be determined.
}
\details{
The procedure (1) calculates the mean scores for the difference between the
observed data \code{x} and the value of mu under the null hypothesis
\code{mu}. The user has the option to enumerate all possible randomizations
or to sample the randomization distribution. In this latter case, the
function \code{sampleRD} is invoked. In addition, it is possible to
produce an approximate 95\% / 99\% confidence interval by interpolation
(useful for simulations).  If confidence limits are needed, the number of
randomizations should be a large enough number (probably 4999 or more). The
upper percentage points (percentages of randomization differences greater
than or equal to the observed \code{x - mu} means) can be determined for a
range of trial values for L and U, which, when subtracted from the \code{x},
just avoid giving a significant difference between the two sample means. The
upper percentage points (percentages of randomization differences greater
than or equal to the observed difference between \code{x} and \code{mu}) can
be determined for a range of trial values for L, and linear interpolation is
used to determine the value of L to be substracted from the \code{x}
measurements producing a difference between \code{x} and \code{mu} that is
on the borderline of being significantly large at about the 2.5\\% or 0.5\\%
level. Analogously, the lower percentage points (the percentages of
randomization differences less than or equal to the observed difference
\code{x - mu} can be determined for some trial values of U. Again, linear
interpolation is used to determine the value of U to be substracted from the
\code{x - mu} measurements producing a difference that is on the borderline
of being significantly small at about the 2.5\\% or 0.5\\%level.

\code{seed} is a way to call the \code{\link[base]{set.seed}} function,
"the recommended way to specify seeds" in random number generation.

The function \code{summary.RT} is used to obtain and print a summary of the
results, and a \code{plot.RT} method is available for displaying the
randomization distribution of mean differences.
}
\examples{

# Example in Manly and Navarro Alberto (2021), Section 1.5
data(cornheight)
Dif <- cornheight[,"Cross"] - cornheight[,"Self"]
corn1s.c <- RT1SAMP(Dif, complete=TRUE) # Complete enumerations
#
corn1s.s <- RT1SAMP(Dif, CI=TRUE) # Sampled randomization distribution, CI for x - mu.


}
\references{
Manly, B.F.J. and Navarro-Alberto, J.A. (2021) Randomization,
Bootstrap and Monte Carlo Methods in Biology. 4th Edition. Chapman and
Hall/ CRC Press.
}
\seealso{
\code{sampleRD}, \code{count}
}
\author{
Jorge Navarro-Alberto
}
