% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RT2SAMP.R
\name{RT2SAMP}
\alias{RT2SAMP}
\title{Randomization Test of the Mean Difference between Two Samples}
\usage{
RT2SAMP(
  x,
  y,
  NRAND = 4999,
  alt = "two",
  CI = FALSE,
  silent = FALSE,
  seed = NULL
)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{a (non-empty) numeric vector of data values.}

\item{NRAND}{numeric; the number of randomizations (permutations).}

\item{alt}{a character string specifying the alternative hypothesis, must
be one of "\code{two}" (two-sided, the default), "\code{greater}" or
"\code{less}".}

\item{CI}{a logical variable indicating whether approximate 95\% and
99\% randomization confidence intervals will be calculated (\code{TRUE})
or not (\code{FALSE}).}

\item{silent}{a logical variable indicating whether calculation results are
printed to the R console (\code{silent = FALSE}). If \code{TRUE} then
calculation results are not printed to the R console (useful for
simulations)}

\item{seed}{a single value, interpreted as an integer, or \code{NULL} (see
"Details").}
}
\value{
The function returns a \code{RT} result object (list)
}
\description{
RT2SAMP carries out a two sample randomization test on the mean difference
between two samples. A randomization confidence interval for the mean
difference between two source populations can also be determined.
}
\details{
The procedure (1) calculates the mean scores for vectors \code{x} and
\code{y}, and the difference \code{D[0]} between these; then (2)
\code{length(x)} and \code{length(y)} observations are randomly reallocated
to the first and second group, respectively, using the \code{\link{sample}}
function. Step (2) is repeated \code{NRAND} times to find the randomization
distribution of \code{D} differences. \code{alt = "greater"} is the
alternative that x has a larger mean than y.

The randomization test allows confidence intervals to be placed on
treatment effects (when \code{CI = TRUE}), as described in Manly and
Navarro (2021, Section 1.4). If confidence limits are needed, \code{NRAND}
should be a large enough number (probably 4999 or more). The upper
percentage points (percentages of randomization differences greater than or
equal to the observed between \code{x} and \code{y} means) can be
determined for a range of trial values for L and U, which, when subtracted
from the \code{x}, just avoid giving a significant difference between the
two sample means. The upper percentage points (percentages of randomization
differences greater than or equal to the observed difference between
\code{x} and \code{y}) can be determined for a range of trial values for L,
and linear interpolation is used to determine the value of L to be
substracted from the \code{x} measurements producing a difference between
\code{x} and \code{y} that is on the borderline of being significantly
large at about the 2.5\% or 0.5\%. Analogously, the lower percentage
points (the percentages of randomization differences less than or equal to
the observed difference between \code{x} and \code{y} means) can be
determined for some trial values of U. Again, linear interpolation is used
to determine the value of U to be substracted from the \code{x}
measurements producing a difference between \code{x} and \code{y} that is
on the borderline of being significantly small at about the 2.5\% or
0.5\% level.

\code{seed} is a way to call the \code{\link[base]{set.seed}} function,
"the recommended way to specify seeds" in random number generation.

The function \code{\link[rbmc]{summary.RT}} is used to obtain and print a summary of the
results, and a \code{plot.RT} method is available for displaying the
randomization distribution of mean differences.
}
\examples{

# Example in Manly and Navarro Alberto (2021), Section 1.1
data(jackals)
male <- jackals$Mand.length[jackals$Sex=="M"]
female <- jackals$Mand.length[jackals$Sex=="F"]
RT2SAMP(male, female, alt="greater")
}
\references{
Manly, B.F.J. and Navarro-Alberto, J.A. (2021) Randomization,
Bootstrap and Monte Carlo Methods in Biology. 4th Edition. Chapman and
Hall/ CRC Press.
}
\seealso{
\code{summary.RT} and the main plotting function \code{plot.RT}
}
\author{
Jorge Navarro-Alberto
}
