% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleRD.R
\name{sampleRD}
\alias{sampleRD}
\title{Sampling a randomization distribution}
\usage{
sampleRD(x, mu = 0, NRAND = 4999)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{mu}{an scalar; it is the value of \code{mu} under the null hypothesis}

\item{NRAND}{numeric; the number of randomizations (permutations).}
}
\value{
The function returns a list containing the value of the \code{sumD}
statistic and the p-values of the sampled randomization distribution for the
three types of alternatives, and the number of randomizations + 1
}
\description{
sampleRD samples the randomization distribution for Fisher or paired sample
test
}
\details{
The procedure calculates first the sumD-statistic for the observed data,
where \code{sumD} stands for the differences \code{x - mu}, and \code{mu} is
a hypothesized value for mu under a null hypothesis (usually set to 0, the
default. Then, the randomization procedure decides to change sign of each
\code{x - mu} or not, after a coin flip. The number of coin flips are
controlled by the parameter \code{NRAND} thus producing the randomization
distribution of \code{sumD} differences.
}
\examples{
data(cornheight)
Dif <- cornheight[,"Cross"] - cornheight[,"Self"]
Results.RD <- sampleRD(Dif)

}
\references{
Manly, B.F.J. and Navarro-Alberto, J.A. (2021) Randomization,
Bootstrap and Monte Carlo Methods in Biology. 4th Edition. Chapman and
Hall/ CRC Press.
}
\seealso{
\code{RT1SAMP}
}
\author{
Jorge Navarro-Alberto
}
