% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.RT.R
\name{summary.RT}
\alias{summary.RT}
\title{Summarize one and two sample randomization tests}
\usage{
\method{summary}{RT}(x)
}
\arguments{
\item{x}{an object of class RT}
}
\description{
Summarize the results produced by RT1SAMP and RT2SAMP
}
\examples{
# Example in Manly and Navarro Alberto (2021), Section 1.5
Dif <- cornheight[,"Cross"] - cornheight[,"Self"]
corn1s.c <- RT1SAMP(Dif, complete=TRUE) # Complete enumeration
summary.RT(corn1s.c)

}
\references{
Manly, B.F.J. and Navarro-Alberto, J.A. (2021) Randomization,
Bootstrap and Monte Carlo Methods in Biology. 4th Edition. Chapman and
Hall/ CRC Press.
}
\seealso{
\code{\link[rbmc]{RT1SAMP}}, \code{\link[rbmc]{RT2SAMP}}, and the
main plotting function \code{\link[rbmc]{plot.RT}}
}
